+++ dart/collision/Option.hpp
@@ -33,8 +33,8 @@
 #ifndef DART_COLLISION_OPTION_HPP_
 #define DART_COLLISION_OPTION_HPP_
 
-#warning "This header has been deprecated in DART 6.1. "\
-  "Please include CollisionOption.hpp intead."
+#pragma message("This header has been deprecated in DART 6.1. "\
+  "Please include CollisionOption.hpp intead.")
 
 #include "dart/collision/CollisionOption.hpp"