+++ examples/bipedStand/MyWindow.hpp
@@ -42,7 +42,7 @@
 
 #include "Controller.hpp"
 
-class MyWindow : public dart::gui::SimWindow {
+class MyWindow : public dart::gui::glut::SimWindow {
 public:
   MyWindow();
   virtual ~MyWindow();