+++ tutorials/tutorialBiped/tutorialBiped.cpp
@@ -45,6 +45,7 @@ using namespace dart::common;
 using namespace dart::dynamics;
 using namespace dart::simulation;
 using namespace dart::gui;
+using namespace dart::gui::glut;
 using namespace dart::io;
 using namespace dart::math;